Feat: added user POST, reorganised function files
