+++ src/controllers/functions/UserFunctions.js
@@ -1,6 +1,6 @@
 // Require specific models so that we can 
 // create functionality involving them.
-const { User } = require('../models/UserModel');
+const { User } = require('../../models/UserModel');
 
 // Model.find({}) returns all documents in a collection.
 async function getAllUsers(){
@@ -12,8 +12,13 @@ async function getUserByUsername(username){
     return await User.findOne({username: username}).exec()
 }
 
+async function createNewUser(data){
+    return await User.create(data).catch((error) => error)
+}
+
 // Export the functions for our routes to use.
 module.exports = {
     getAllUsers,
-    getUserByUsername
+    getUserByUsername,
+    createNewUser
 }
